Emit patchable funcs under -enable-jit-support
This is a squash commit of:

Teach IRGen to emit patchable functions

Add an option to the IRGen module will cause
it to add the patchable-function attribute to
the list of default function attributes, which
will in turn tell LLVM to emit patchable
functions.

Emit 32 bytes of prefix for patchable functions

Needed to support effective function patching and useful for storing
runtime metadata for/with the function.

Set IRGen patchable function for JIT support

Tell IRGen to emit patchable functions when
-enable-jit-support is in effect.